convert project to Vite
