+++ src/App.tsx
@@ -1,7 +1,7 @@
 import { Vol3dViewer } from '@janelia/web-vol-viewer';
 import './App.css';
 import React from 'react';
-import CloudViewerUI from './CloudViewerUI.jsx';
+import CloudViewerUI from './CloudViewerUI';
 
 
 function App() {
