+++ tsconfig.json
@@ -1,0 +1,8 @@
+{
+    "compilerOptions": {
+        "lib": ["dom", "dom.iterable", "esnext"],
+        "target": "ESNext",
+        "types": ["vite/client"],
+        "isolatedModules": true,
+    },
+ }
